+++ actix-server/src/lib.rs
@@ -3,8 +3,10 @@
 mod accept;
 mod builder;
 mod config;
+mod counter;
 mod server;
 mod services;
+pub mod ssl;
 mod worker;
 
 pub use self::builder::ServerBuilder;